pyproject.toml
[tool.poetry]
name = "myFastAPI"
version = "0.1.0"
description = "Basic setup with FastAPI+Pydantic.v2+SQLAlchemy.v2+Postgres"
authors = ["Yuri Chiucconi <yuri.chiucconi@exact-lab.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
fastapi = {extras = ["all"], version = "^0.104.1"}
pydantic = "^2.4.2"
sqlalchemy = "^2.0.23"
asyncpg = "^0.28.0"
psycopg2 = "^2.9.9"
alembic = "^1.12.1"
devtools = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
